requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
faiss-cpu
langchain_community
langchain_google_genai